# docs/plans/artifacts/openapi.snapshot.json
{
  "components": {
    "schemas": {
...
      "ComplianceTransactionRequest": {
        "description": "Check a transaction against compliance rules.",
        "properties": {
          "rules": {
            "description": "Optional rule overrides. If empty, default rules are used.",
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Rules",
            "type": "array"
          },
          "transaction": {
            "additionalProperties": true,
            "description": "Transaction record: {id, timestamp, user_id, order_id, symbol, side, quantity, price, venue, strategy_id}",
            "title": "Transaction",
            "type": "object"
          }
        },
        "required": [
          "transaction"
        ],
        "title": "ComplianceTransactionRequest",
        "type": "object"
      },
...
      "FactorCombineRequest": {
        "description": "Combine multiple factors with optional timing.",
        "properties": {
          "benchmark": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": "XU100",
            "title": "Benchmark"
          },
          "end_date": {
            "default": "2025-12-31",
            "title": "End Date",
            "type": "string"
          },
          "max_position_weight": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Position Weight"
          },
          "method": {
            "default": "custom",
            "description": "Weighting scheme: custom, equal, risk_parity, mean_variance, min_variance",
            "title": "Method",
            "type": "string"
          },
          "rebalance_frequency": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rebalance Frequency"
          },
          "signals": {
            "description": "List of {name, weight, ...params} factor specs",
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "minItems": 1,
            "title": "Signals",
            "type": "array"
          },
          "start_date": {
            "default": "2020-01-01",
            "title": "Start Date",
            "type": "string"
          },
          "timing_enabled": {
            "default": false,
            "title": "Timing Enabled",
            "type": "boolean"
          },
          "timing_lookback": {
            "default": 63,
            "minimum": 5.0,
            "title": "Timing Lookback",
            "type": "integer"
          },
          "timing_threshold": {
            "default": 0.0,
            "title": "Timing Threshold",
            "type": "number"
          },
          "top_n": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Top N"
          },
          "weights": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "number"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Weights"
          }
        },
        "required": [
          "signals"
        ],
        "title": "FactorCombineRequest",
        "type": "object"
      },
      "FactorSnapshotRequest": {
        "description": "Cross-sectional signal scores request.",
        "properties": {
          "buy_threshold": {
            "default": 0.3,
            "title": "Buy Threshold",
            "type": "number"
          },
          "custom_symbols": {
            "items": {
              "type": "string"
            },
            "title": "Custom Symbols",
            "type": "array"
          },
          "focus_symbol": {
            "default": "",
            "title": "Focus Symbol",
            "type": "string"
          },
          "indicator_params": {
            "additionalProperties": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Indicator Params",
            "type": "object"
          },
          "indicators": {
            "items": {
              "type": "string"
            },
            "title": "Indicators",
            "type": "array"
          },
          "interval": {
            "default": "1d",
            "title": "Interval",
            "type": "string"
          },
          "max_symbols": {
            "default": 100,
            "maximum": 500.0,
            "minimum": 1.0,
            "title": "Max Symbols",
            "type": "integer"
          },
          "period": {
            "default": "1y",
            "title": "Period",
            "type": "string"
          },
          "sell_threshold": {
            "default": -0.3,
            "title": "Sell Threshold",
            "type": "number"
          },
          "top_n": {
            "default": 50,
            "maximum": 500.0,
            "minimum": 1.0,
            "title": "Top N",
            "type": "integer"
          },
          "universe": {
            "default": "XU100",
            "title": "Universe",
            "type": "string"
          }
        },
        "title": "FactorSnapshotRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PositionLimitsRequest": {
        "description": "Check position limit breaches.",
        "properties": {
          "positions": {
            "description": "List of {name, value, limit, ...} position entries",
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "minItems": 1,
            "title": "Positions",
            "type": "array"
          }
        },
        "required": [
          "positions"
        ],
        "title": "PositionLimitsRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "BIST Quant API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/compliance/activity-anomalies": {
      "post": {
        "description": "Detect anomalous user activity from event logs.",
        "operationId": "compliance_activity_anomalies_api_compliance_activity_anomalies_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "title": "Payload",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Compliance Activity Anomalies Api Compliance Activity Anomalies Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Compliance Activity Anomalies",
        "tags": [
          "compliance"
        ]
      }
    },
    "/api/compliance/check": {
      "post": {
        "description": "Check a transaction against compliance rules.",
        "operationId": "compliance_check_api_compliance_check_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ComplianceTransactionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Compliance Check Api Compliance Check Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Compliance Check",
        "tags": [
          "compliance"
        ]
      }
    },
    "/api/compliance/position-limits": {
      "post": {
        "description": "Check for position limit breaches.",
        "operationId": "compliance_position_limits_api_compliance_position_limits_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PositionLimitsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Compliance Position Limits Api Compliance Position Limits Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Compliance Position Limits",
        "tags": [
          "compliance"
        ]
      }
    },
...
    "/api/factors/combine": {
      "post": {
        "description": "Combine multiple factors with weighting and optional timing.",
        "operationId": "factors_combine_api_factors_combine_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FactorCombineRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Factors Combine Api Factors Combine Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Factors Combine",
        "tags": [
          "factors"
        ]
      }
    },
    "/api/factors/snapshot": {
      "post": {
        "description": "Run cross-sectional signal scoring snapshot.",
        "operationId": "factors_snapshot_api_factors_snapshot_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FactorSnapshotRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Factors Snapshot Api Factors Snapshot Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Factors Snapshot",
        "tags": [
          "factors"
        ]
      }
    },
    "/api/factors/{name}": {
      "get": {
        "description": "Get detailed info for a single signal/factor.",
        "operationId": "factors_detail_api_factors__name__get",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Factors Detail Api Factors  Name  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Factors Detail",
        "tags": [
          "factors"
        ]
      }
    },
...
  }
}